requirements.txt
weaviate-client
streamlit
pillow
requests
